requirements.txt
google-api-python-client>=2.70.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
tqdm>=4.65.0

# Optional (for abstractive summarizer)
transformers>=4.40.0
torch>=2.2.0   # only if you want transformer-based summarization locally

fastapi
uvicorn

sentencepiece

nltk
beautifulsoup4
lxml

python-multipart
reportlab

pydantic
anyio
typing-extensions
annotated-types